requirements.txt
requests
psycopg2-binary
Flask
pymongo
psycopg2-binary
gunicorn
requests
